permissions module

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -3,6 +3,7 @@ const jwt = require('jsonwebtoken');
 const { randomBytes } = require('crypto');
 const { promisify } = require('util');
 const { niceEmail, transport } = require('../mail');
+const { hasPermission } = require('../utils');
 
 const Mutations = {
   async createItem(parent, args, ctx, info) {
@@ -14,7 +15,7 @@ const Mutations = {
     const item = await ctx.db.mutation.createItem(
       {
         data: {
-          // This is how to creat a relationship between the item and the user
+          // This is how to create a relationship between the item and the user
           user: {
             connect: {
               id: ctx.request.userId,
@@ -42,8 +43,16 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
     // Find item
-    const item = await ctx.db.query.item({ where }, `{ id title}`);
-    // TODO: check if they own that item
+    const item = await ctx.db.query.item({ where }, `{ id title user { id }}`);
+    // check if they own that item, or have the permissions
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ['ADMIN', 'ITEMDELETE'].includes(permission),
+    );
+
+    if (!ownsItem && !hasPermissions) {
+      throw new Error('You do not have permissions to do that!');
+    }
 
     // Delete it
     return ctx.db.mutation.deleteItem({ where }, info);
@@ -188,6 +197,41 @@ const Mutations = {
 
     // 8. Return the new user
     return updatedUser;
+  },
+
+  async updatePermissions(parent, args, ctx, info) {
+    // Check if they are logged in
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in!');
+    }
+
+    // Query the current user
+    const currentUser = await ctx.db.query.user(
+      {
+        where: {
+          id: ctx.request.userId,
+        },
+      },
+      info,
+    );
+
+    // Check if they have permissions to do this
+    hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE']);
+
+    // Update the permissions
+    return ctx.db.mutation.updateUser(
+      {
+        data: {
+          permissions: {
+            set: args.permissions,
+          },
+        },
+        where: {
+          id: args.userId,
+        },
+      },
+      info,
+    );
   },
 };
